Router add, Small additions.
